Fix crawler.py to handle empty tracks list
When the tweet crawler encounters an empty tracks list, the program will
now raise an exception instead of defaulting to the "python" track, as
was the case before.

In addition to clarifying the program behavior, this will make it easier
to debug any edge cases having to do with an empty streaming tracks list.