Add a check for auth config keys in crawler.py
While the os module will automatically raise an exception if no
environment variables exist, this commit adds checks so that the
provided config values are "valid" (i.e. non-empty).

This ensures that (1) the user has provided an .env file and
(2) the user has filled out the .env file with the proper values.