Fix runtime errors due to invalid PWD variable
Before this patch, docker-compose would suffer a fatal runtime error
if the user did not define a PWD environment variable. Given that the
PWD environment variable is not set up by default, this would result
in a program crash without clear feedback from docker-compose.

The simplest solution is to swap out the PWD variable with the 'dot'
symbol for the current working directory, without touching the user's
terminal environment variables. While less flexible, this solution
results in a working program without requiring any action on the part
of users.

Closes #8.