Add the full_tweet field to Twitter status objects
The standard tweet objects returned by the Twitter API follow
an inconsistent format, in that they may or may not contain the
full body text of a tweet within the 'text' field. In particular,
long-form tweets (> 140 chars) will contain truncated data in the
'text' field while containing the full data within the
'extended_tweet.full_text' field.

Given that most applications are interested primarily in the full
text of a tweet, this commit extracts the full text for each tweet
and stores it within a new field called 'full_tweet'. In addition to
making it easier to access the full tweet, this change makes full-text
search on tweets much simpler and more robust.