Update crawler.py to handle client stalling
When attempting to stream highly popular topics via the Twitter
Streaming API (e.g. "america", "world cup", etc.), the tweet
crawler would raise a fatal exception due to an inability to
consume all of the tweets produced by the API.

This commit simply restarts the crawler when it gets hit by the
exception, so that it can recover from stall failures. It is
important to note that this solution will result in some tweets
being dropped, due to the "abort and restart" strategy being used.
A more robust solution would make use of some kind of queuing
technology (like RabbitMQ).

Closes #9.