Split main Dockerfile into multiple sub-files
Early on in the project, development was focused on
creating a crawler that consumed exclusively from the
'filter' Twitter streaming API. This crawler was
specialized to only allow streaming based on topics
of interest (i.e. tracks).

With this commit, the main Dockerfile has been split
into several sub-files, each of which can be used to
define containers that run geocrawler.py, sampler.py,
and crawler.py on their own. In this manner, we can
instantiate multiple specialized crawlers, each in
their own container.