Update databases.py to handle duplicates
Prior versions of the MongoDB driver code would
crash due to an exception raised when trying
to insert duplicate data into a collection which
has a unique index defined on it. In particular,
defining a unique index over Twitter IDs helps
to cut down on duplicate records, but will result
in an exception when the client tries to insert
duplicate tweets.

This commit allows the code to try insertion for
each value provided to add_tweets and will ignore
any records containing duplicate Tweet IDs.