Add ability to convert raw datestrings to datetimes
Before this commit, raw datestrings received from the Twitter API were
stored as-is in the tweets database. Although this simplified the
storage of tweets, this made it difficult for clients to run range
queries on the date fields (created_at, user.created_at, etc.).

In order to resolve this, this patch updates the MongoDB driver so that
it first converts any datestrings into native datetime objects before
storing the data into the database.