Add support for Redis-based task queue
Prior to this commit, the crawler would directly write data to
the tweets database. This resulted in significant slowdown due
to the I/O overhead involved in writing to disk-based MongoDB.

After this commit, the crawler will make use of an intermediate
Redis-based task queue to schedule asynchronous jobs for writing
to MongoDB. In addition to decoupling the database from the crawler,
this update results in significant improvements to crawler efficiency.