Remove logic for handling stall warnings
Before this commit, the crawler contained logic for
handling stall warnings sent by the Twitter API, which
forced the crawler to restart and reconnect. Given that
the slow crawler runtime was to blame for this, the
previous commit, which added support for a Redis-based
task queue, has rendered this code more or less unnecessary.

Although we could retain the handler code to help with slow
connections, its removal results in a smaller codebase and,
hence, an overall more maintainable crawler project.