Update docker-compose.yaml to support task queue
This commit simply builds on the previous two commits
to update docker-compose.yaml with support for the new
"crawler-redis-mongo" workflow for the tweet crawling
process.

In particular, after applying this commit, running the
"docker-compose run" command will create a crawler that
writes to MongoDB using a Redis aync worker, rather than
via direct I/O operations on the database, as before.