Fix keyboard interrupt handling logic in crawler
In previous versions of the crawler, hitting Ctrl^C would
abort the crawling process and print out an informative
message to the user. However, any tweets which were not
persisted to the database would be lost forever.

This patch fixes this bug by calling the exception
handler defined on the main stream listener attached to
the crawler object. In particular, the exception handler
will persist any collected tweets to Redis before
shutting down gracefully.