Add crawler support for daily indexed collections
Prior to this patch, all tweets collected by the crawler
were to be written to a single, monolithic database collection.
Although simple in the short-term, this storage scheme is
potentially very costly in the long-term. As such, support has
been added to allow the crawler to segment the collected tweet
data into daily "buckets".

In addition to splitting the collected data on a daily basis,
this commit updates the 'add_tweet' method in databases.py to
create text and multikey indexes for faster searches on the
full tweet text, hashtags, and user mentions.