Fix crawler crash issues due to output on STDOUT
In certain cases, running the crawler with output stream
set to STDOUT will crash the crawler due to stall warnings
received from the Twitter API. Given that this is an
intermittent and rare issue, the compose file has been
updated to automatically restart the crawler if it crashes
due to such stall warnings.

Note that the "restart" setting will only apply if the user
runs the "pycrawler" service via the "docker-compose run"
command.