Update README.md with build and development info
Prior to this update, the README.md file was mainly
intended to address issues from a user's point-of-view.
As such, there was very little information about
how to use the tweet-crawler project from a developer's
point-of-view.

This patch adds new information for building the project's
various Docker images, as well instructions for performing
specific tasks using the Tweet Crawler system.